+++ src/bootstrap/index.ts
@@ -60,6 +60,14 @@ function _ready(config: BalmConfig): BalmConfig {
     );
   }
 
+  // Set HMR flag
+  const HMR_ENV =
+    config.useDefaults &&
+    config.env.isDev &&
+    !config.env.inSSR &&
+    !config.env.isMP;
+  config.server.useHMR = HMR_ENV && !config.scripts.ie8;
+
   return config;
 }